Closes #23. Remove mini oozie

+++ iis-core/src/test/java/eu/dnetlib/iis/core/AbstractOozieWorkflowTestCase.java
@@ -98,7 +98,9 @@ public abstract class AbstractOozieWorkflowTestCase {
 	
 	@After
 	public void cleanup() throws IOException {
-		FileUtils.deleteDirectory(tempDir);
+		if (tempDir != null) {
+			FileUtils.deleteDirectory(tempDir);
+		}
 	}
 	
 	/**
